Backend/python-worker/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
pypdf
